+++ inst/include/dataobject.h
@@ -66,6 +66,7 @@ public:
   unsigned int get_idata_row(const double ID);  
   void locate_tran();
   void get_records(recstack& a, int NID, int neq, unsigned int& obscount, unsigned int& evcount, bool obsonly,bool debug);
+  void get_records_pred(recstack& a, int NID, int neq, unsigned int& obscount, unsigned int& evcount, bool obsonly,bool debug);
   void check_idcol(dataobject& data);
   double get_value(const int row, const int col) const {return Data(row,col);}
   double get_id_value(const int row) const {return Data(row,Idcol);}