+++ inst/models/pred1.cpp
@@ -1,0 +1,8 @@
+$PROB
+An example model expressed in closed form
+$PARAM B = -1, beta0 = 100, beta1 = 0.1
+$OMEGA 2 0.3
+$PRED
+double beta0i = beta0 + ETA(1);
+double beta1i = beta1*exp(ETA(2));
+capture Y = beta0i + beta1i*B;