+++ src/housemodel-mread-source.cpp
@@ -1,4 +1,4 @@
-// Source MD5: 3e4121cfa1e1b6db274a660ef635b5c8
+// Source MD5: 509e24de6401c4c7d8c72c3487b55e52
 
 
 // FIXED:
@@ -38,12 +38,17 @@ typedef double localdouble;
 #define __CONFIGFUN___ _model_housemodel_config__
 #define __REGISTERFUN___ R_init_housemodel
 #define _nEQ 3
-#define _nPAR 13
+#define _nPAR 14
 #define N_GUT 1
+#define N_CENT 2
 #define F_GUT _F_[0]
+#define F_CENT _F_[1]
 #define ALAG_GUT _ALAG_[0]
+#define ALAG_CENT _ALAG_[1]
 #define R_GUT _R_[0]
+#define R_CENT _R_[1]
 #define D_GUT _D_[0]
+#define D_CENT _D_[1]
 #define GUT _A_[0]
 #define CENT _A_[1]
 #define RESP _A_[2]
@@ -57,15 +62,16 @@ typedef double localdouble;
 #define VC _THETA_[1]
 #define KA _THETA_[2]
 #define F1 _THETA_[3]
-#define WT _THETA_[4]
-#define SEX _THETA_[5]
-#define WTCL _THETA_[6]
-#define WTVC _THETA_[7]
-#define SEXCL _THETA_[8]
-#define SEXVC _THETA_[9]
-#define KIN _THETA_[10]
-#define KOUT _THETA_[11]
-#define IC50 _THETA_[12]
+#define D1 _THETA_[4]
+#define WT _THETA_[5]
+#define SEX _THETA_[6]
+#define WTCL _THETA_[7]
+#define WTVC _THETA_[8]
+#define SEXCL _THETA_[9]
+#define SEXVC _THETA_[10]
+#define KIN _THETA_[11]
+#define KOUT _THETA_[12]
+#define IC50 _THETA_[13]
 #define ECL _xETA(1)
 #define EVC _xETA(2)
 #define EKA _xETA(3)
@@ -79,6 +85,7 @@ __END_config__
 // MAIN CODE BLOCK:
 __BEGIN_main__
 F_GUT = F1;
+D_CENT = D1;
 CLi   = exp(log(CL)   + WTCL*log(WT/70) + log(SEXCL)*SEX + ECL);
 VCi   = exp(log(VC)   + WTVC*log(WT/70) + log(SEXVC)*SEX + EVC);
 KAi   = exp(log(KA)   + EKA);